third_party/blink/web_tests/platform/generic/virtual/text-antialias/tab-min-size-expected.txt

Tab characters should never be zero-width.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS oneTabLineWidth is not equal to twoTabLineWidth.
PASS successfullyParsed is true

TEST COMPLETE

	word
		word
